#148 - test: isbn13으로 좋아요 여부 확인 테스트 코드 작성
