+++ src/test/java/cotato/bookitlist/book/controller/BookLikeControllerTest.java
@@ -13,11 +13,10 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @Transactional
 @SpringBootTest
@@ -114,6 +113,71 @@ class BookLikeControllerTest {
                         .content(objectMapper.writeValueAsBytes(request))
                 )
                 .andExpect(status().isNotFound())
+                .andDo(print())
+        ;
+    }
+
+    @Test
+    @WithCustomMockUser
+    @DisplayName("찜한 책 목록을 조회한다.")
+    void givenLoginMember_whenGettingLikePosts_thenReturnBookListResponse() throws Exception {
+        //given
+
+        //when & then
+        mockMvc.perform(get("/books/likes/all"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.totalResults").value(2))
+        ;
+    }
+
+    @Test
+    @DisplayName("로그인 없이 찜한 책 목록을 조회하면 에러를 반환한다.")
+    void givenNonLoginMember_whenGettingLikePosts_thenReturnErrorResponse() throws Exception {
+        //given
+
+        //when & then
+        mockMvc.perform(get("/books/likes/all"))
+                .andExpect(status().isUnauthorized())
+        ;
+    }
+
+    @Test
+    @WithCustomMockUser
+    @DisplayName("로그인한 유저가 책을 좋아요 했는지 확인한다.")
+    void givenLoginMember_whenGettingLiked_thenReturnBookLikeResponse() throws Exception {
+        //given
+        String isbn13 = "9788931514810";
+
+        //when & then
+        mockMvc.perform(get("/books/likes")
+                        .param("isbn13", isbn13))
+                .andExpect(status().isOk())
+                .andDo(print())
+        ;
+    }
+
+    @Test
+    @WithCustomMockUser
+    @DisplayName("로그인한 유저가 isbn13 없이 책을 좋아요를 확인하면 에러를 반환한다.")
+    void givenNonIsbn13_whenGettingLiked_thenReturnErrorResponse() throws Exception {
+        //given
+
+        //when & then
+        mockMvc.perform(get("/books/likes"))
+                .andExpect(status().isBadRequest())
+                .andDo(print())
+        ;
+    }
+
+    @Test
+    @DisplayName("로그인안한 유저가 책을 좋아요 했는지 확인하면 에러를 반환한다.")
+    void givenMember_whenGettingLiked_thenReturnErrorResponse() throws Exception {
+        //given
+        String isbn13 = "9788931514810";
+
+        //when & then
+        mockMvc.perform(get("/books/likes"))
+                .andExpect(status().isUnauthorized())
                 .andDo(print())
         ;
     }
